Integrate live reload into toolchain.
