+++ package.json
@@ -11,20 +11,6 @@
   "bugs": {
     "url": "http://github.com/mozilla/browser.html/issues/"
   },
-  "browserify": {
-    "transform": [
-      [
-        "babelify",
-        {
-          "optional": [
-            "spec.protoToAssign",
-            "runtime"
-          ],
-          "blacklist": []
-        }
-      ]
-    ]
-  },
   "scripts": {
     "start": "gulp develop",
     "server": "gulp server",
@@ -42,16 +28,19 @@
     "typed-immutable": "0.0.7"
   },
   "devDependencies": {
+    "babel-core": "5.8.23",
     "babel-runtime": "5.8.20",
     "babelify": "6.1.3",
     "browserify": "11.0.1",
     "ecstatic": "0.8.0",
     "graphene-marionette-runner": "0.0.5",
     "gulp": "3.9.0",
+    "gulp-sequence": "0.4.1",
     "gulp-sourcemaps": "1.5.2",
     "gulp-uglify": "^1.2.0",
     "gulp-util": "^3.0.6",
     "jscs": "1.8.1",
+    "livereactload": "0.7.0",
     "marionette-client": "1.7.5",
     "mocha": "2.2.4",
     "vinyl-buffer": "1.0.0",
